fix: Soft #712 max-lines ratchet + rustfmt after rebase
Extract workspace_add and slug/remove helpers so main.rs and
managed_workspace.rs stay within baseline. Bump conformance baseline
to match main (1678). Apply cargo fmt.
