+++ rust/alera-cli/src/main.rs
@@ -24,6 +24,7 @@ mod managed_workspace;
 mod managed_workspace_handoff;
 #[cfg(test)]
 mod managed_workspace_removal_tests;
+mod managed_workspace_slug;
 mod mobile_access;
 mod native_credential_entry;
 mod netbird;
@@ -37,6 +38,7 @@ mod project_management;
 mod pty_job_bootstrap;
 mod remote_managed_workspace;
 mod remote_managed_workspace_git;
+mod remote_managed_workspace_remove;
 mod remote_workspace_files;
 mod runtime_archive;
 mod runtime_clear;
@@ -50,6 +52,7 @@ mod tab_record_factory;
 mod tailscale;
 mod terminal_alias_commands;
 mod terminal_host;
+mod workspace_add;
 mod workspace_context;
 mod workspace_handoff;
 mod workspace_pinning;
@@ -80,7 +83,7 @@ use crate::cli::{
     CascadePreviewArgs, Cli, Command, IdArgs, ProjectAction, ProjectAddArgs, ProjectCommand,
     ProjectKindArg, RuntimeDirArgs, SshAuthKindArg, SshTargetAction, SshTargetAddArgs,
     SshTargetBootstrapArgs, SshTargetBootstrapPlanArgs, SshTargetCommand, SshTargetStatusArgs,
-    TabAction, TabCommand, WorkspaceAction, WorkspaceAddArgs, WorkspaceCommand,
+    TabAction, TabCommand, WorkspaceAction, WorkspaceCommand,
 };
 use crate::cli::{MobileAction, MobileCommand, MobileDevicesAction, MobilePairingAction};
 use crate::cli::{TerminalAction, TerminalCommand};
@@ -347,31 +350,7 @@ async fn run_workspace_command(command: WorkspaceCommand) -> i32 {
             return workspace_handoff::run_hand_on(runtime, args, json_output).await;
         }
         WorkspaceAction::Add(args) => {
-            let remote = crate::ssh_remote::is_remote_host_id(args.host_id.as_deref());
-            let payload = match workspace_add_payload(args) {
-                Ok(payload) => payload,
-                Err(error) => return print_error(error),
-            };
-            let client = if remote {
-                RuntimeHostRpcClient::connect_or_start_with_required_capability(
-                    &runtime_dir(&runtime),
-                    crate::terminal_host::protocol::RUNTIME_HOST_REMOTE_SSH_WORKSPACES_CAPABILITY,
-                )
-                .await
-            } else {
-                runtime_host_required(&runtime).await
-            };
-            let value: Value = match client {
-                Ok(mut client) => match client
-                    .request_value("workspace.createManaged", &payload)
-                    .await
-                {
-                    Ok(value) => value,
-                    Err(error) => return print_error(error),
-                },
-                Err(error) => return print_error(error),
-            };
-            print_value(&value, json_output, "workspace created");
+            return workspace_add::run(runtime, args, json_output).await;
         }
         WorkspaceAction::Setup(args) => {
             let client = match runtime_host_required(&runtime).await {
@@ -993,7 +972,9 @@ where
     store_operation(open_store(args).await?).await
 }
 
-async fn runtime_host_required(args: &RuntimeDirArgs) -> anyhow::Result<RuntimeHostRpcClient> {
+pub(crate) async fn runtime_host_required(
+    args: &RuntimeDirArgs,
+) -> anyhow::Result<RuntimeHostRpcClient> {
     RuntimeHostRpcClient::connect_or_start(&runtime_dir(args)).await
 }
 
@@ -1229,30 +1210,6 @@ fn host_accessible_path(value: String, current_dir: &Path) -> PathBuf {
     }
 }
 
-fn workspace_add_payload(args: WorkspaceAddArgs) -> anyhow::Result<Value> {
-    let remote = crate::ssh_remote::is_remote_host_id(args.host_id.as_deref());
-    Ok(json!({
-        "id": args.id,
-        "projectId": args.project_id,
-        "name": args.name,
-        "branch": args.branch,
-        "sourceBranch": args.source_branch,
-        "reuseExistingBranch": args.reuse_existing_branch,
-        "workspaceRoot": if remote {
-            normalized_workspace_path_value(args.workspace_root.as_deref().unwrap_or(""))
-        } else {
-            host_accessible_optional_string_path(args.workspace_root)?
-        },
-        "path": if remote {
-            normalized_workspace_path_value(args.path.as_deref().unwrap_or(""))
-        } else {
-            host_accessible_optional_string_path(args.path)?
-        },
-        "parentWorkspaceId": args.parent_workspace_id,
-        "hostId": args.host_id,
-    }))
-}
-
 pub(crate) fn host_accessible_optional_string_path(
     value: Option<String>,
 ) -> anyhow::Result<Option<String>> {
@@ -1270,7 +1227,7 @@ pub(crate) fn host_accessible_optional_string_path(
     ))
 }
 
-fn normalized_workspace_path_value(value: &str) -> Option<String> {
+pub(crate) fn normalized_workspace_path_value(value: &str) -> Option<String> {
     let trimmed = value.trim();
     if trimmed.is_empty() {
         None
